Test: 리뷰, 알림 삭제 API 관련 테스트 코드 수정 #54
Related to #54
