+++ tests/reviews/selectors/reviews/test_get_review_with_concept_queryset_by_photo_booth_brand_id_order_by_created_at_desc.py
@@ -43,15 +43,6 @@ class TestGetReviewWithConceptQuerysetByPhotoBoothBrandIdOrderByCreatedAtDesc:
 
         assert review_with_concept_queryset.count() == 0
 
-    def test_get_review_with_concept_queryset_by_photo_booth_brand_id_order_by_created_at_desc_fail_deleted_review(self, deleted_review):
-        review_with_concept_queryset = (
-            self.review_selector.get_review_with_concept_queryset_by_photo_booth_brand_id_order_by_created_at_desc(
-                deleted_review.photo_booth.photo_booth_brand.id
-            )
-        )
-
-        assert review_with_concept_queryset.count() == 0
-
     def test_get_review_with_concept_queryset_by_photo_booth_brand_id_order_by_created_at_desc_fail_private_review(self, private_review):
         review_with_concept_queryset = (
             self.review_selector.get_review_with_concept_queryset_by_photo_booth_brand_id_order_by_created_at_desc(
