+++ tests/reviews/selectors/reviews/test_get_review_with_photo_booth_queryset_by_user_like.py
@@ -25,13 +25,6 @@ class TestGetReviewWithPhotoBoothQuerysetByUserLike:
 
         assert review_queryset_with_photo_booth.count() == 0
 
-    def test_get_review_with_photo_booth_queryset_by_user_like_fail_deleted_review(self, deleted_review, valid_user):
-        deleted_review.user_review_like_logs.add(valid_user)
-
-        review_queryset_with_photo_booth = self.review_selector.get_review_with_photo_booth_queryset_by_user_like(valid_user.id)
-
-        assert review_queryset_with_photo_booth.count() == 0
-
     def test_get_review_with_photo_booth_queryset_by_user_like_fail_private_review(self, private_review, valid_user):
         private_review.user_review_like_logs.add(valid_user)
